LPS-60025 move also this method to the Base
